Changed tool to generate indices rather than be provided with them
Summary: Ref T10845.

Test Plan: `./test.sh poplibs`

Reviewers: tomm

Reviewed By: tomm

Maniphest Tasks: T10845

Differential Revision: https://phabricator.sourcevertex.net/D13979

+++ include/poplibs_test/Embedding.hpp
@@ -6,11 +6,11 @@ namespace poplibs_test {
 namespace embedding {
 
 void multiSlice(const boost::multi_array<double, 2> &embeddingMatrix,
-                const std::vector<long unsigned> &indices,
+                const std::vector<unsigned> &indices,
                 boost::multi_array<double, 2> &result);
 
 void multiUpdateAdd(const boost::multi_array<double, 2> &deltas,
-                    const std::vector<long unsigned> &indices,
+                    const std::vector<unsigned> &indices,
                     const double scale,
                     boost::multi_array<double, 2> &embeddingMatrix);
