+++ lib/poplibs_test/Embedding.cpp
@@ -6,7 +6,7 @@ namespace poplibs_test {
 namespace embedding {
 
 void multiSlice(const boost::multi_array<double, 2> &embeddingMatrix,
-                const std::vector<long unsigned> &indices,
+                const std::vector<unsigned> &indices,
                 boost::multi_array<double, 2> &result) {
   const auto size = embeddingMatrix.shape()[1];
   if (size != result.shape()[1]) {
@@ -30,7 +30,7 @@ void multiSlice(const boost::multi_array<double, 2> &embeddingMatrix,
 }
 
 void multiUpdateAdd(const boost::multi_array<double, 2> &deltas,
-                    const std::vector<long unsigned> &indices,
+                    const std::vector<unsigned> &indices,
                     const double scale,
                     boost::multi_array<double, 2> &embeddingMatrix) {
   const auto size = deltas.shape()[1];
